Fix session bug and lack of 403 handling. Closes #154.
Add 403 handling to expire current session and show 403. This
fixes a small bug: if someone is logged in and changes the application
oauth settings then they would see an octobouncer cat with no message
or help as to what was happening. This bug also appeared if the user
revoked access to play and wanted to readd it.

Added 403.html to public and styling for it and also iphone version.
Nothing fancy. Steal Octobouncer for our own gains.