improve Firefox search field size
In Firefox, the search field was so large that it jutted below the black bar it sits in.

This has no effect on the size in Safari and Chrome. In Firefox, the search field would look even better at 15px, but 15px bumps the field down too small in Safari and Chrome.