Cache album artwork.
Play has to shell out to get album art for music. This makes play cache the
images so it doesn't have to shell out.

Album artwork is saved to play/public/art. The filename is generated by
creating a SHA hash of the Artist Name and Album Name and is a png image.

When the play app is started it creates a new thread to precache the artwork
for all the music the MPD library. When there is no album art or the art
hasn't been cached yet the art-placeholder.png is sent in place.