Paginate search
Since Song itself isn't backed by a database, we have to manually construct
the WillPaginate collection. That's all fine and dandy, but we also can't load
all of the Song records from disk, either, so instead run the fast MPD query and
then construct a WIllPaginate::Collection with just that sliver of Songs.