Allow dots in specific URL's. Fixes #342
Rails by convention parses the dot in the last part of the URL (after the slash) as a format. This conflicts with song/artist/album titles which include a dot in the name, incorrectly parsing the last part of the name as a dot.

Using constraints, we get around this issue by allowing the params to be anything but a slash.